fix: DB 구성에 따라 class의 field별 type 수정

+++ backend/src/models/projectRanking.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, Float, ForeignKey
 from sqlalchemy.orm import relationship
 from api.db import Base
 
@@ -8,7 +8,7 @@ class ProjectRanking(Base):
 
     project_id = Column("PROJECT_ID", Integer, ForeignKey("PROJECT.PROJECT_ID"), primary_key=True)
     freelancer_id = Column("FREELANCER_ID", Integer, ForeignKey("FREELANCER.FREELANCER_ID"), primary_key=True)
-    matching_score = Column("MATCHING_SCORE", Integer, nullable=False)
+    matching_score = Column("MATCHING_SCORE", Float, nullable=False)
     applied = Column("APPLIED", Integer, nullable=False)
 
     # 관계 정의
